Rename selection_gestures to tap_and_drag_gestures
