+++ packages/flutter/lib/src/widgets/text_selection.dart
@@ -24,7 +24,7 @@ import 'gesture_detector.dart';
 import 'magnifier.dart';
 import 'overlay.dart';
 import 'scrollable.dart';
-import 'selection_gestures.dart';
+import 'tap_and_drag_gestures.dart';
 import 'tap_region.dart';
 import 'ticker_provider.dart';
 import 'transitions.dart';
