+++ packages/flutter/lib/widgets.dart
@@ -118,7 +118,6 @@ export 'src/widgets/scrollable.dart';
 export 'src/widgets/scrollbar.dart';
 export 'src/widgets/selectable_region.dart';
 export 'src/widgets/selection_container.dart';
-export 'src/widgets/selection_gestures.dart';
 export 'src/widgets/semantics_debugger.dart';
 export 'src/widgets/service_extensions.dart';
 export 'src/widgets/shared_app_data.dart';
@@ -136,6 +135,7 @@ export 'src/widgets/spacer.dart';
 export 'src/widgets/spell_check.dart';
 export 'src/widgets/status_transitions.dart';
 export 'src/widgets/table.dart';
+export 'src/widgets/tap_and_drag_gestures.dart';
 export 'src/widgets/tap_region.dart';
 export 'src/widgets/text.dart';
 export 'src/widgets/text_editing_intents.dart';
